feat: Admin을 Product 목록 서비스에서 분리하고 cursor secret 제거
